requirements.txt
astroid==2.3.3
colorama==0.4.3
Django==2.2.9
django-crispy-forms==1.8.1
django-filter==2.2.0
djangorestframework==3.11.0
django-webpack-loader==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
pip==20.0.2
pylint==2.4.4
pytz==2019.3
setuptools==41.2.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
